rename app -> geo
