+++ geo/geo/broker.py
@@ -7,7 +7,7 @@ import asyncio
 import logging
 from uuid import uuid4
 from pybroker import BrokerBase
-from app.errors import MissingReplyFieldError
+from geo.errors import MissingReplyFieldError
 
 
 log = logging.getLogger(__name__)
