+++ geo/geo/errors.py
