+++ geo/geo/state/database.py
@@ -10,11 +10,11 @@ from sqlalchemy import create_engine
 from sqlalchemy import select
 from sqlalchemy import func
 from sqlalchemy.orm import Session
-from app.geonames import CityRow
-from app.state.schema import Base
-from app.state.schema import UpdateTracker
-from app.state.schema import Name
-from app.state.schema import Place
+from geo.geonames import CityRow
+from geo.state.schema import Base
+from geo.state.schema import UpdateTracker
+from geo.state.schema import Name
+from geo.state.schema import Place
 
 log = logging.getLogger(__name__)
