+++ geo/tests/test_database.py
@@ -3,11 +3,11 @@ from random import random
 import pytest
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-from app.geonames import CityRow
-from app.state.database import Database
-from app.state.schema import Name
-from app.state.schema import Place
-from app.state.schema import UpdateTracker
+from geo.geonames import CityRow
+from geo.state.database import Database
+from geo.state.schema import Name
+from geo.state.schema import Place
+from geo.state.schema import UpdateTracker
 
 @pytest.fixture
 def config():
